Add a fallback reading time calculation
The wordCount property is only available in the Reader v1.2 API.
When hitting the v2 API (e.g. Discover), this information won't
be available.

To work around this, we are going to parse the content string and
split it by space. However, note that this is just an approximation
since there are HTML tags in the content.